2 սինի պատճառով ֆայլը կիսում եմ
